processTweets: fix for 4c4b8d1
The original patch didn't fix the problem that List.get(int) is used which
only makes sense when the list also is java.util.RandomAccess. In case of
a LinkedList it would've been better to just move forward/backward from
the 'current' item at 'pos'.

This patch makes the underlying ArrayList explicit, which may also be a
good idea since List.removeAll() is an optional operation and we cannot
necessarily assume that Lemmatization.lemmatize() will indeed return a
mutable list (there's no comment about it in that API's description). See
also the following patch.