Lemmatization: reduce useless creation of temporaries and clarify API.
'forms' is an unmodified temporary that's used exactly once and therefore
can die.

Since lemmatize() returns a List<String>, for this change it can be
assumed that code calling this function doesn't make use of internals of
the implementation of this method (namely: that it's actually a *mutable*
*Array*List). So it's possible to change the implementation, allowing to
get rid of another "intermediary copy" to the freshly created ArrayList
'lemmas'. Maybe callers don't want to modify our result and just iterate
over the items. Then this copy would have been completely unnecessary.