processTweets: remove implicit assumption: lemmatize() returns a RandomAccess List
The algorithm in processTweets() relies on the (so far implicit) assumption
that the List retrieved from Lemmatization.lemmatize() is
java.util.RandomAccess since it makes extensive use of list.get(int).
In case of a LinkedList it would've been better to just move forward/backward
from the 'current' item at 'pos'.

This patch makes the underlying ArrayList explicit, which may also be a
good idea since List.removeAll() is an optional operation and we cannot
necessarily assume that Lemmatization.lemmatize() will indeed return a
mutable list (there's no comment about it in that API's description).

Removal of this assumption also enables the following patch.